Add ability to set BUILD_SHORT=headers to include header files without building gcc/binutils (useful with tinycc).

+++ mini-native.sh
@@ -73,7 +73,7 @@ cleanup busybox
 
 if [ ! -z "${BUILD_SHORT}" ]
 then
-  rm -rf "${TOOLS}"/include
+  [ "$BUILD_SHORT" != "headers" ] && rm -rf "${TOOLS}"/include
 else
 
 # Build and install native binutils
